+++ argopy/data_fetchers/argovis_data.py
@@ -18,7 +18,7 @@ import warnings
 
 from argopy.stores import httpstore
 from argopy.options import OPTIONS
-from argopy.utilities import list_standard_variables, format_oneline, is_box, Chunker
+from argopy.utilities import list_standard_variables, format_oneline, Chunker
 from argopy.plotters import open_dashboard
 
 access_points = ['wmo', 'box']
@@ -274,8 +274,6 @@ class Fetch_wmo(ArgovisDataFetcher):
             CYC : int, np.array(int), list(int)
                 The cycle numbers to load.
         """
-        if isinstance(WMO, int):
-            WMO = [WMO]  # Make sure we deal with a list
         if isinstance(CYC, int):
             CYC = np.array(
                 (CYC,), dtype="int"
@@ -349,7 +347,6 @@ class Fetch_box(ArgovisDataFetcher):
             start = end - pd.DateOffset(months=1)
             box.append(start.strftime("%Y-%m-%d"))
             box.append(end.strftime("%Y-%m-%d"))
-        is_box(box)
         self.BOX = box
 
         self.definition = "?"