+++ argopy/data_fetchers/erddap_index.py
@@ -217,8 +217,6 @@ class Fetch_wmo(ErddapArgoIndexFetcher):
             WMO : list(int)
                 The list of WMOs to load all Argo data for.
         """
-        if isinstance(WMO, int):
-            WMO = [WMO]  # Make sure we deal with a list
         self.WMO = WMO
         self.definition = 'Ifremer erddap Argo Index fetcher for floats'
         return self
@@ -259,8 +257,6 @@ class Fetch_box(ErddapArgoIndexFetcher):
             # Use all time line:
             box.append('1900-01-01')
             box.append('2100-12-31')
-        elif len(box) != 6:
-            raise ValueError('Box must have 4 or 6 elements : [lon_min, lon_max, lat_min, lat_max, datim_min, datim_max] ')
         self.BOX = box
         self.definition = 'Ifremer erddap Argo Index fetcher for a space/time region'